Remove cruft [CI SKIP]
