+++ include/boost/math/special_functions/prime_sieve.hpp
@@ -5,222 +5,41 @@
 // (See accompanying file LICENSE_1_0.txt
 // or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#ifndef BOOST_MATH_SPECIAL_FUNCTIONS_PRIME_SIEVE_HPP
-#define BOOST_MATH_SPECIAL_FUNCTIONS_PRIME_SIEVE_HPP
+#ifndef BOOST_MATH_SPECIAL_FUNCTIONS_PRIME_SEIVE_ITER_HPP
+#define BOOST_MATH_SPECIAL_FUNCTIONS_PRIME_SEIVE_ITER_HPP
 
-#include <boost/math/special_functions/prime.hpp>
-#include <boost/math/special_functions/interval_sieve.hpp>
+#include <boost/math/special_functions/detail/linear_prime_sieve.hpp>
+#include <boost/math/special_functions/detail/interval_prime_sieve.hpp>
+#include <boost/math/special_functions/detail/small_primes.hpp>
 #include <boost/math/special_functions/prime_approximation.hpp>
-#include <boost/assert.hpp>
-#include <vector>
-#include <iterator>
-#include <cmath>
-#include <thread>
-#include <memory>
-#include <future>
-#include <numeric>
-#include <algorithm>
 #include <execution>
+#include <vector>
+#include <future>
+#include <thread>
+#include <limits>
 #include <type_traits>
+#include <iterator>
+#include <cstdint>
 
-namespace boost::math { namespace detail
+namespace boost::math::detail::prime_sieve
 {
-// https://mathworld.wolfram.com/SieveofEratosthenes.html
-// https://www.cs.utexas.edu/users/misra/scannedPdf.dir/linearSieve.pdf
-template<class Integer, class Container>
-void linear_sieve(Integer upper_bound, Container &resultant_primes)
+inline std::size_t L1D_SIZE {32'768};
+
+template<typename Integer, typename OutputIterator>
+decltype(auto) sequential_segmented_sieve(const Integer lower_bound, const Integer upper_bound, OutputIterator resultant_primes)
 {
-    std::size_t least_divisors_size{static_cast<std::size_t>(upper_bound + 1)};
-    std::unique_ptr<Integer[]> least_divisors{new Integer[least_divisors_size]{0}};
+    const Integer interval {static_cast<Integer>(L1D_SIZE * 8)};
+    Integer current_lower_bound {lower_bound};
+    Integer current_upper_bound {current_lower_bound + interval};
 
-    for (std::size_t i{2}; i < upper_bound; ++i)
-    {
-        if (least_divisors[i] == 0)
-        {
-            least_divisors[i] = i;
-            resultant_primes.emplace_back(i);
-        }
-
-        for (std::size_t j{}; j < resultant_primes.size() && i * resultant_primes[j] <= upper_bound && 
-             resultant_primes[j] <= least_divisors[i] && j < least_divisors_size; ++j)
-        {
-            least_divisors[i * static_cast<std::size_t>(resultant_primes[j])] = resultant_primes[j];
-        }
-    }
-}
-
-// 4096 is where benchmarked performance of linear_sieve begins to diverge
-template<class Integer>
-const Integer linear_sieve_limit = Integer(4096); // Constexpr does not work with boost::multiprecision types
-
-template<class Integer, class PrimeContainer, class Container>
-void mask_sieve(Integer lower_bound, Integer upper_bound, const PrimeContainer& primes, Container &resultant_primes)
-{
-    Integer limit {static_cast<Integer>(std::floor(std::sqrt(static_cast<double>(upper_bound)))) + 1};
-
-    std::size_t primes_size {};
-    auto it{primes.begin()};
-    while(it != primes.end() && *it < limit)
-    {
-        ++primes_size;
-        ++it;
-    }
-
-    const std::size_t n {static_cast<std::size_t>(upper_bound - lower_bound + 1)};
-    std::unique_ptr<bool[]> is_prime {new bool[n]};
-    memset(is_prime.get(), true, sizeof(*is_prime.get()) * (n));
-     
-    // Enable use of thread pool, not SIMD compatible
-    std::for_each(std::execution::par, primes.begin(), it, [&is_prime, lower_bound, upper_bound](auto prime){
-        for(Integer j {std::max(static_cast<std::size_t>(prime * prime), static_cast<std::size_t>((lower_bound + prime - 1) / prime * prime))}; 
-            j < upper_bound; j += prime)
-        {
-            is_prime[static_cast<std::size_t>(j - lower_bound)] = false;
-        }
-    });
-
-    if(lower_bound == 1)
-    {
-        is_prime[0] = false;
-    }
-
-    for(Integer i{lower_bound}; i <= upper_bound; ++i)
-    {
-        if(is_prime[static_cast<std::size_t>(i - lower_bound)])
-        {
-            resultant_primes.emplace_back(i);
-        }
-    }
-}
-
-template<class Integer, class Container>
-void mask_sieve(Integer lower_bound, Integer upper_bound, Container &resultant_primes)
-{
-    auto limit{std::floor(std::sqrt(static_cast<double>(upper_bound))) + 1};
-    std::vector<Integer> primes {};
-    primes.reserve(limit / std::log(limit));
-
-    boost::math::detail::linear_sieve(limit, primes);
-
-    boost::math::detail::mask_sieve(lower_bound, upper_bound, primes, resultant_primes);
-}
-
-template<class Integer, class Container>
-constexpr void prime_table(std::size_t min_index, Integer upper_bound, Container &resultant_primes)
-{
-    std::size_t current_index {min_index};
-    Integer current_prime {2};
-
-    while(current_prime < upper_bound)
-    {
-        resultant_primes.emplace_back(current_prime);
-        ++current_index;
-        current_prime = prime(current_index);
-    }
-}
-
-template<class Integer, class Container>
-constexpr void prime_table(Integer upper_bound, Container &resultant_primes)
-{
-    prime_table(0, upper_bound, resultant_primes);
-}
-
-template<class Integer, class PrimesContainer, class Container>
-void segmented_sieve(Integer lower_bound, Integer upper_bound, const PrimesContainer &primes, Container &resultant_primes)
-{
-    const Integer L1_SIZE {32768};
-    const Integer interval {L1_SIZE * 8};
-    Integer current_lower_bound{lower_bound};
-    Integer current_upper_bound{current_lower_bound + interval};
-
-    if(current_upper_bound > upper_bound)
+    if (current_upper_bound > upper_bound)
     {
         current_upper_bound = upper_bound;
     }
 
     std::size_t ranges {static_cast<std::size_t>((upper_bound - lower_bound) / interval)};
 
-    std::vector<std::vector<Integer>> prime_vectors(ranges + 1);
-    std::vector<std::future<void>> future_manager(ranges);
-
-    auto primes_in_range {static_cast<std::size_t>(static_cast<double>(current_upper_bound) / std::log(static_cast<double>(current_upper_bound)) -
-                          static_cast<double>(current_lower_bound) / std::log(static_cast<double>(current_lower_bound)))};
-
-    for(std::size_t i {}; i < ranges; ++i)
-    {
-        prime_vectors[i].reserve(primes_in_range);
-
-        future_manager.emplace_back(std::async(std::launch::async, [current_lower_bound, current_upper_bound, &primes, &prime_vectors, i]{
-            boost::math::detail::IntervalSieve sieve(current_lower_bound, current_upper_bound, primes, prime_vectors[i]);
-        }));
-
-        current_lower_bound = current_upper_bound;
-        current_upper_bound += interval;
-    }
-
-    prime_vectors[ranges].reserve(primes_in_range);
-    future_manager.emplace_back(std::async(std::launch::async, [current_lower_bound, upper_bound, &primes, &prime_vectors]{
-        boost::math::detail::IntervalSieve sieve(current_lower_bound, upper_bound, primes, prime_vectors.back());
-    }));
-
-    for(auto &&future : future_manager)
-    {
-        if(future.valid())
-        {
-            future.get();
-        }
-    }
-
-    for(auto &v : prime_vectors)
-    {
-        resultant_primes.insert(resultant_primes.end(), v.begin(), v.end());
-    }
-}
-
-template<class Integer, class Container>
-void segmented_sieve(Integer lower_bound, Integer upper_bound, Container &resultant_primes)
-{
-    using boost::math::detail::linear_sieve_limit;
-    
-    Integer limit{static_cast<Integer>(std::floor(std::sqrt(static_cast<double>(upper_bound)))) + 1};
-    std::vector<Integer> primes {};
-    primes.reserve(static_cast<double>(limit) / std::log(static_cast<double>(limit)));
-
-    // Prepare for max value so you do not have to calculate this again
-    if(limit < linear_sieve_limit<Integer>)
-    {
-        boost::math::detail::linear_sieve(static_cast<Integer>(limit), primes);
-    }
-
-    else
-    {
-        boost::math::detail::linear_sieve(linear_sieve_limit<Integer>, primes);
-        boost::math::detail::segmented_sieve(linear_sieve_limit<Integer>, limit, primes, primes);
-    }
-
-    boost::math::detail::segmented_sieve(lower_bound, upper_bound, primes, resultant_primes);
-}
-
-template<class Integer, class Container>
-void sequential_segmented_sieve(Integer lower_bound, Integer upper_bound, Container &resultant_primes)
-{
-    const Integer L1_SIZE {32768};
-    const Integer interval {L1_SIZE * 8};
-    Integer current_lower_bound{lower_bound};
-    Integer current_upper_bound{current_lower_bound + interval};
-
-    if(current_upper_bound > upper_bound)
-    {
-        current_upper_bound = upper_bound;
-    }
-
-    std::size_t ranges {static_cast<std::size_t>((upper_bound - lower_bound) / interval)};
-
-    boost::math::detail::IntervalSieve sieve(current_lower_bound, current_upper_bound, resultant_primes, resultant_primes);
-    if(ranges == 0)
-    {
-        return;
-    }
+    IntervalSieve sieve(current_lower_bound, current_upper_bound, resultant_primes);
 
     for(std::size_t i {}; i < ranges; ++i)
     {
@@ -230,235 +49,224 @@ void sequential_segmented_sieve(Integer lower_bound, Integer upper_bound, Contai
         {
             current_upper_bound = upper_bound;
         }
-        sieve.NewRange(current_lower_bound, current_upper_bound, resultant_primes);
+        resultant_primes = sieve.NewRange(current_lower_bound, current_upper_bound);
     }
+    
+    return resultant_primes;
+}
+
+template<typename Integer, typename OutputIterator>
+decltype(auto) segmented_sieve(const Integer lower_bound, const Integer upper_bound, OutputIterator resultant_primes)
+{
+    const auto num_threads {std::thread::hardware_concurrency() > 0 ? std::thread::hardware_concurrency() : 2u};
+    const Integer thread_range {(upper_bound - lower_bound) / static_cast<Integer>(num_threads)};
+
+    std::vector<std::vector<Integer>> prime_vectors(num_threads);
+    std::vector<std::future<void>> future_manager;
+    future_manager.reserve(num_threads);
+
+    Integer current_lower_bound {lower_bound};
+    Integer current_upper_bound {current_lower_bound + thread_range};
+
+    for(std::size_t i {}; i < num_threads - 1; ++i)
+    {
+        prime_vectors[i].resize(static_cast<std::size_t>(prime_approximation(current_lower_bound, current_upper_bound)));
+
+        future_manager.emplace_back(std::async(std::launch::async, [current_lower_bound, current_upper_bound, &prime_vectors, i]{
+            sequential_segmented_sieve(current_lower_bound, current_upper_bound, prime_vectors[i].begin());
+        }));
+
+        current_lower_bound = current_upper_bound;
+        current_upper_bound += thread_range;
+    }
+
+    prime_vectors.back().resize(static_cast<std::size_t>(prime_approximation(current_lower_bound, upper_bound)));
+    future_manager.emplace_back(std::async(std::launch::async, [current_lower_bound, upper_bound, &prime_vectors]{
+            sequential_segmented_sieve(current_lower_bound, upper_bound, prime_vectors.back().begin());
+    }));
+
+    std::size_t i {};
+    for(auto&& future : future_manager)
+    {
+        future.get(); // Blocks to maintain proper sorting
+        
+        for(auto& val : prime_vectors[i])
+        {
+            *resultant_primes++ = std::move(val);
+        }
+
+        ++i;
+    }
+    
+    return resultant_primes;
+}
+
+template<typename ExecutionPolicy, typename Integer, typename OutputIterator>
+decltype(auto) prime_sieve_iter_impl(ExecutionPolicy&& policy, const Integer upper_bound, OutputIterator resultant_primes)
+{
+    if (upper_bound == 2)
+    {
+        return resultant_primes;
+    }
+
+    else if (upper_bound <= small_prime_limit<Integer>())
+    {
+        small_primes(upper_bound, resultant_primes);
+        return resultant_primes;
+    }
+    
+    resultant_primes = small_primes(small_prime_limit<Integer>(), resultant_primes);
+
+    if constexpr (std::is_same_v<std::remove_reference_t<decltype(policy)>, decltype(std::execution::seq)> 
+                  #if __cpp_lib_execution > 201900
+                  || std::is_same_v<std::remove_reference_t<decltype(policy)>, decltype(std::execution::unseq)>
+                  #endif
+                  )
+    {
+        resultant_primes = sequential_segmented_sieve(small_prime_limit<Integer>(), upper_bound, resultant_primes);
+    }
+
+    else
+    {
+        resultant_primes = segmented_sieve(small_prime_limit<Integer>(), upper_bound, resultant_primes);
+    }
+    
+    return resultant_primes;
+}
+
+template<typename Integer, typename OutputIterator>
+inline decltype(auto) prime_sieve_iter_impl(const Integer upper_bound, OutputIterator resultant_primes)
+{
+    return prime_sieve_iter_impl(std::execution::seq, upper_bound, resultant_primes);
+}
+
+template<typename ExecutionPolicy, typename Integer, typename OutputIterator>
+decltype(auto) prime_range_iter_impl(ExecutionPolicy&& policy, const Integer lower_bound, const Integer upper_bound, OutputIterator resultant_primes)
+{
+    if(lower_bound <= small_prime_limit<Integer>() || upper_bound <= small_prime_limit<Integer>())
+    {
+        std::vector<Integer> primes;
+        primes.resize(static_cast<std::size_t>(prime_approximation(upper_bound)));
+        small_primes(upper_bound, primes.begin());
+
+        auto it {primes.begin()};
+        while(*it < lower_bound)
+        {
+            ++it;
+        }
+
+        while(it != primes.end() && *it <= upper_bound)
+        {
+            *resultant_primes++ = std::move(*it++);
+        }
+    }
+
+    if(lower_bound <= small_prime_limit<Integer>() && upper_bound > small_prime_limit<Integer>())
+    {
+        if constexpr (std::is_same_v<std::remove_reference_t<decltype(policy)>, decltype(std::execution::seq)> 
+                      #if __cpp_lib_execution > 201900
+                      || std::is_same_v<std::remove_reference_t<decltype(policy)>, decltype(std::execution::unseq)>
+                      #endif
+                     )
+        {
+            resultant_primes = detail::prime_sieve::sequential_segmented_sieve(small_prime_limit<Integer>(), upper_bound, resultant_primes);
+        }
+
+        else
+        {
+            resultant_primes = detail::prime_sieve::segmented_sieve(small_prime_limit<Integer>(), upper_bound, resultant_primes);
+        }
+    }
+
+    else
+    {
+        if constexpr (std::is_same_v<std::remove_reference_t<decltype(policy)>, decltype(std::execution::seq)> 
+                      #if __cpp_lib_execution > 201900
+                      || std::is_same_v<std::remove_reference_t<decltype(policy)>, decltype(std::execution::unseq)>
+                      #endif
+                     )
+        {
+            resultant_primes = detail::prime_sieve::sequential_segmented_sieve(lower_bound, upper_bound, resultant_primes);
+        }
+
+        else
+        {
+            resultant_primes = detail::prime_sieve::segmented_sieve(lower_bound, upper_bound, resultant_primes);
+        }
+    }
+
+    return resultant_primes;
+}
+
+template<typename Integer, typename OutputIterator>
+inline decltype(auto) prime_range_iter_impl(const Integer lower_bound, const Integer upper_bound, OutputIterator resultant_primes)
+{
+    return prime_range_iter_impl(std::execution::seq, lower_bound, upper_bound, resultant_primes);
 }
 
 // SFINAE for dual interface
-template <typename T>
+template<typename T>
 class is_container
 {
-    typedef char yes;
+private:
+    using yes = char; 
     struct no { char x[2]; };
 
-    template <typename U> static yes test( decltype(&U::size) );
-    template <typename U> static no test(...);    
+    template<typename U>  
+    static yes test( decltype(&U::size) );
+    
+    template<typename U> 
+    static no test(...);    
 
 public:
-    enum { value = sizeof(test<T>(0)) == sizeof(char) };
+    enum { type = sizeof(test<T>(0)) == sizeof(char) };
 };
 
-template<class ExecutionPolicy, class Integer, class Container>
-void prime_sieve_impl(ExecutionPolicy&& policy, Integer upper_bound, Container &primes)
-{
-    using boost::math::detail::linear_sieve_limit;
-
-    if(upper_bound == 2)
-    {
-        return;
-    }
-
-    if(upper_bound <= linear_sieve_limit<Integer>)
-    {
-        boost::math::detail::linear_sieve(static_cast<Integer>(upper_bound), primes);
-    }
-
-    else if constexpr (std::is_same_v<std::remove_reference_t<decltype(policy)>, decltype(std::execution::seq)> || 
-                       #if __cpp_lib_execution > 201900
-                       std::is_same_v<std::remove_reference_t<decltype(policy)>, decltype(std::execution::unseq)>
-                       #endif
-                       )
-    {
-        boost::math::detail::linear_sieve(linear_sieve_limit<Integer>, primes);
-        boost::math::detail::sequential_segmented_sieve(linear_sieve_limit<Integer>, upper_bound, primes);
-    }
-
-    else
-    {
-        std::vector<Integer> small_primes {};
-        small_primes.reserve(1028);
-
-        std::thread t1([&small_primes] {
-            boost::math::detail::linear_sieve(static_cast<Integer>(linear_sieve_limit<Integer> * 2), small_primes);
-        });
-        std::thread t2([upper_bound, &primes] {
-            boost::math::detail::segmented_sieve(static_cast<Integer>(linear_sieve_limit<Integer> * 2), upper_bound, primes);
-        });
-
-        t1.join();
-        t2.join();
-        primes.insert(primes.begin(), small_primes.begin(), small_primes.end());
-    }
+// Add C++14/17esqe interface to is_container
+template<typename T>
+static constexpr auto is_container_t = is_container<T>::type;
 }
 
-template<class Integer, class Container>
-void prime_sieve_impl(Integer upper_bound, Container &primes)
+namespace boost::math
 {
-    prime_sieve_impl(std::execution::seq, upper_bound, primes);
-}
-
-template<typename ExecutionPolicy, typename Integer, typename OutputIterator>
-inline decltype(auto) prime_sieve_wrapper(ExecutionPolicy&& policy, Integer upper_bound, OutputIterator resultant_primes)
+template<typename ExecutionPolicy, typename Integer, typename T>
+inline decltype(auto) prime_sieve(ExecutionPolicy&& exec, Integer upper_bound, T output)
 {
-    std::vector<Integer> primes;
-    prime_reserve(upper_bound, primes);
-    prime_sieve_impl(policy, upper_bound, primes);
-
-    return std::move(primes.begin(), primes.end(), resultant_primes);
-}
-
-template<typename Integer, typename OutputIterator>
-inline decltype(auto) prime_sieve_wrapper(Integer upper_bound, OutputIterator resultant_primes)
-{
-    return prime_sieve_wrapper(std::execution::seq, upper_bound, resultant_primes);
-}
-
-template<class ExecutionPolicy, class Integer, class Container>
-void prime_range_impl(ExecutionPolicy&& policy, Integer lower_bound, Integer upper_bound, Container &primes)
-{
-    using boost::math::detail::linear_sieve_limit;
-    Integer limit {static_cast<Integer>(std::floor(std::sqrt(static_cast<double>(upper_bound)))) + 1};
-
-    if(upper_bound == 2)
+    if constexpr (detail::prime_sieve::is_container_t<std::remove_pointer_t<T>>)
     {
-        return;
-    }
-
-    if(upper_bound <= linear_sieve_limit<Integer>)
-    {
-        boost::math::detail::linear_sieve(static_cast<Integer>(upper_bound), primes);
-    }
-
-    else if constexpr (std::is_same_v<std::remove_reference_t<decltype(policy)>, decltype(std::execution::seq)> || 
-                       #if __cpp_lib_execution > 201900
-                       std::is_same_v<std::remove_reference_t<decltype(policy)>, decltype(std::execution::unseq)>
-                       #endif
-                       )
-    {
-        if(limit <= linear_sieve_limit<Integer>)
-        {   
-            boost::math::detail::linear_sieve(limit, primes);
-            
-            if(lower_bound <= limit)
+        if constexpr (std::is_integral_v<Integer>)
+        {
+            return detail::prime_sieve::prime_sieve_iter_impl(exec, upper_bound, std::begin(*output));
+        }
+        else
+        {
+            if (upper_bound < static_cast<Integer>(std::numeric_limits<std::size_t>::max()))
             {
-                boost::math::detail::sequential_segmented_sieve(limit, upper_bound, primes);
+                return detail::prime_sieve::prime_sieve_iter_impl(exec, static_cast<std::size_t>(upper_bound), std::begin(*output));
             }
             else
             {
-                boost::math::detail::sequential_segmented_sieve(lower_bound, upper_bound, primes);
+                return detail::prime_sieve::prime_sieve_iter_impl(exec, upper_bound, std::begin(*output));
             }
-            
-        }
-
-        else
-        {
-            boost::math::detail::linear_sieve(linear_sieve_limit<Integer>, primes);
-            boost::math::detail::sequential_segmented_sieve(linear_sieve_limit<Integer>, limit, primes);
-            boost::math::detail::sequential_segmented_sieve(lower_bound, upper_bound, primes);
         }
     }
-
     else
     {
-        std::vector<Integer> small_primes {};
-
-        if(limit <= static_cast<Integer>(linear_sieve_limit<Integer> * 2))
+        if constexpr (std::is_integral_v<Integer>)
         {
-            small_primes.reserve(1028);
-
-            std::thread t1([limit, &small_primes] {
-                boost::math::detail::linear_sieve(limit, small_primes);
-            });
-            
-            std::thread t2([lower_bound, limit, upper_bound, &primes] {
-                if(lower_bound <= limit)
-                {
-                    boost::math::detail::segmented_sieve(limit, upper_bound, primes);
-                }
-                else
-                {
-                    boost::math::detail::segmented_sieve(lower_bound, upper_bound, primes);
-                }
-                
-            });
-
-            t1.join();
-            t2.join();
-
-            primes.insert(primes.begin(), small_primes.begin(), small_primes.end());
+            return detail::prime_sieve::prime_sieve_iter_impl(exec, upper_bound, output);
         }
-
         else
         {
-            boost::math::prime_reserve(limit, small_primes);
-
-            std::thread t1([&small_primes] {
-                boost::math::detail::linear_sieve(static_cast<Integer>(linear_sieve_limit<Integer> * 2), small_primes);
-            });
-
-            std::thread t2([limit, &primes] {
-                boost::math::detail::segmented_sieve(static_cast<Integer>(linear_sieve_limit<Integer> * 2), limit, primes);
-            });
-
-            t1.join();
-            t2.join();
-
-            primes.insert(primes.begin(), small_primes.begin(), small_primes.end());
-
-            boost::math::detail::segmented_sieve(lower_bound, upper_bound, primes);
+            if (upper_bound < static_cast<Integer>(std::numeric_limits<std::size_t>::max()))
+            {
+                return detail::prime_sieve::prime_sieve_iter_impl(exec, static_cast<std::size_t>(upper_bound), output);
+            }
+            else
+            {
+                return detail::prime_sieve::prime_sieve_iter_impl(exec, upper_bound, output);
+            }
         }
-    }
-
-    auto it{primes.begin()};
-    while(*it < lower_bound && it != primes.end())
-    {
-        ++it;
-    }
-
-    primes.erase(primes.begin(), it);
-}
-
-template<class Integer, class Container>
-inline void prime_range_impl(Integer lower_bound, Integer upper_bound, Container &primes)
-{
-    prime_range_impl(std::execution::seq, lower_bound, upper_bound, primes);
-}
-
-template<typename ExecutionPolicy, typename Integer, typename OutputIterator>
-inline decltype(auto) prime_range_wrapper(ExecutionPolicy&& policy, Integer lower_bound, Integer upper_bound, OutputIterator resultant_primes)
-{
-    std::vector<Integer> primes;
-    prime_reserve(lower_bound, upper_bound, primes);
-    prime_range_impl(policy, lower_bound, upper_bound, primes);
-
-    return std::move(primes.begin(), primes.end(), resultant_primes);
-}
-
-template<typename Integer, typename OutputIterator>
-inline decltype(auto) prime_range_wrapper(Integer lower_bound, Integer upper_bound, OutputIterator resultant_primes)
-{
-    return prime_range_wrapper(std::execution::seq, lower_bound, upper_bound, resultant_primes);
-}
-
-} // End namespace detail
-
-
-
-
-
-template<typename ExecutionPolicy, typename Integer, typename T>
-inline decltype(auto) prime_sieve(ExecutionPolicy&& policy, Integer upper_bound, T output)
-{
-    if constexpr (detail::is_container<std::remove_pointer_t<T>>::value)
-    {
-        detail::prime_sieve_impl(policy, upper_bound, *output);
-        return;
-    }
-
-    else
-    {
-        detail::prime_sieve_wrapper(policy, upper_bound, output);
-        return output;
     }
 }
 
@@ -469,18 +277,44 @@ inline decltype(auto) prime_sieve(Integer upper_bound, T output)
 }
 
 template<typename ExecutionPolicy, typename Integer, typename T>
-inline decltype(auto) prime_range(ExecutionPolicy&& policy, Integer lower_bound, Integer upper_bound, T output)
+inline decltype(auto) prime_range(ExecutionPolicy&& exec, Integer upper_bound, Integer lower_bound, T output)
 {
-    if constexpr (detail::is_container<std::remove_pointer_t<T>>::value)
+    if constexpr (detail::prime_sieve::is_container_t<std::remove_pointer_t<T>>)
     {
-        detail::prime_range_impl(policy, lower_bound, upper_bound, *output);
-        return;
+        if constexpr (std::is_integral_v<Integer>)
+        {
+            return detail::prime_sieve::prime_range_iter_impl(exec, lower_bound, upper_bound, std::begin(*output));
+        }
+        else
+        {
+            if (upper_bound < static_cast<Integer>(std::numeric_limits<std::size_t>::max()))
+            {
+                return detail::prime_sieve::prime_range_iter_impl(exec, static_cast<std::size_t>(lower_bound), 
+                                                                  static_cast<std::size_t>(upper_bound), std::begin(*output));
+            }
+            else
+            {
+                return detail::prime_sieve::prime_range_iter_impl(exec, lower_bound, upper_bound, std::begin(*output));
+            }
+        }
     }
-
     else
     {
-        detail::prime_range_wrapper(policy, lower_bound, upper_bound, output);
-        return output;
+        if constexpr (std::is_integral_v<Integer>)
+        {
+            return detail::prime_sieve::prime_range_iter_impl(exec, lower_bound, upper_bound, output);
+        }
+        else
+        {
+            if (upper_bound < static_cast<Integer>(std::numeric_limits<std::size_t>::max()))
+            {
+                return detail::prime_sieve::prime_range_iter_impl(exec, static_cast<std::size_t>(lower_bound), static_cast<std::size_t>(upper_bound), output);
+            }
+            else
+            {
+                return detail::prime_sieve::prime_range_iter_impl(exec, lower_bound, upper_bound, output);
+            }
+        }
     }
 }
 
@@ -489,6 +323,12 @@ inline decltype(auto) prime_range(Integer lower_bound, Integer upper_bound, T ou
 {
     return prime_range(std::execution::seq, lower_bound, upper_bound, output);
 }
+
+template<typename Integer>
+inline void set_l1d_size(const Integer size)
+{
+    detail::prime_sieve::L1D_SIZE = static_cast<std::size_t>(size);
+}
 }
 
-#endif //BOOST_MATH_SPECIAL_FUNCTIONS_PRIME_SIEVE_HPP
+#endif // BOOST_MATH_SPECIAL_FUNCTIONS_PRIME_SEIVE_ITER_HPP
